WIP key agreement outputs to an intermediate_secret
Make key agreement output to psa_intermediate_secret_t instead of a
psa_crypto_generator_t. A psa_intermediate_secret_t is conceptually a
handle to a small byte string which can be trimmed and exported or
passed as input to a KDF.

Define a few utility functions for the new object type.