ecp_modp: remove a sanity check in an inner loop
This sanity check is not very useful: it detects either a coding error
which would be caught by functional tests anyway, or a runtime glitch
which isn't any more likely than many glitches that go untested.
Remove it from ecp_modp, which is an inner loop so performance is
important.