Add a test component with malloc(0) returning NULL
Exercise the library functions with malloc(0) returning NULL. Make
this a separte job with UBSan (and ASan) to detect potential. Other
test components jobs either use the system malloc which returns
non-NULL on Linux and FreeBSD, or the memory_buffer_alloc malloc which
returns NULL but does not give as useful feedback with ASan (because
the whole heap is a single C object).