CTR_DRBG: during initial seeding, grab entropy for a nonce
CTR_DRBG always grabbed MBEDTLS_CTR_DRBG_ENTROPY_LEN during the
initial seeding. In some configurations, in particular when using
SHA-256 for entropy, this was a bottleneck on the security strength:
with only 32 bytes of entropy, there was not enough entropy for both
the entropy input and a nonce per NIST specifications. This required a
workaround where the user had to pass a nonce through the `custom`
parameter of mbedtls_ctr_drbg_seed() to achieve the desired security
strength in some cases.

To make things simple, always make a second call to the f_entropy
function during the initial seeding, to grab a nonce.

Update the code, the documentation and the tests accordingly.

For users who want to pass a nonce explicitly, provide a new function
mbedtls_ctr_drbg_seed_with_nonce(). We use this to process the NIST
test vectors.

For programs that were formerly using mbedtls_ctr_drbg_seed(), here
is the evolution path:
* Most programs use CTR_DRBG together with an entropy source
  (typically provided by the library's own entropy module) in order to
  build a random generator. These programs do not need to change and
  will get the benefits of improved security and more evident
  compliance to the NIST standard.
* Programs that use CTR_DRBG and passed an explicit nonce together
  with the personalization string as the `custom` argument to
  mbedtls_ctr_drbg_seed() can continue to do so.
* Programs that use CTR_DRBG as a deterministic generator with
  an f_entropy function that produces known output must switch to
  calling mbedtls_ctr_drbg_seed_with_nonce() with nonce_len=0.