CTR_DRBG: during initial seeding, grab entropy for a nonce
CTR_DRBG always grabbed MBEDTLS_CTR_DRBG_ENTROPY_LEN during the
initial seeding. In some configurations, in particular when using
SHA-256 for entropy, this was a bottleneck on the security strength:
with only 32 bytes of entropy, there was not enough entropy for both
the entropy input and a nonce per NIST specifications. This required a
workaround where the user had to pass a nonce through the `custom`
parameter of mbedtls_ctr_drbg_seed() to achieve the desired security
strength in some cases.

To make things simple, always make a second call to the f_entropy
function during the initial seeding, to grab a nonce.

Update the code, the documentation and the tests accordingly.

Modify the internal function mbedtls_ctr_drbg_set_entropy_len() to
grab the desired amount of entropy when running NIST test vectors. As
this function was not used from any other module (it was only exposed
for the sake of the test suite), this does not break the ABI as far as
library users are concerned.